requirements.txt
discord.py
db-sqlite3
requests
interactions
discord-py-slash-command
discord_components
